Bump major version.
This is an API break, so we have to bump. Fortunately it's not breaking the API
we are using, just restoring the compatibility to Sabre's 2.1 API. Once we
confirm that this new version is working just fine with current stable Horde,
we can simply bump the maximum allowed version in the applications' package.xml
files.